Fix multiple non-singleton credentials in inner conjunction and add send_email
The idea to group all attributes in a single inner conjunction and all initiator-specific
attributes in another distinct inner conjunction was based on that making it possible to add
disjunction labels to indicate which attributes are requested for what. However, since
disjunction labels are not supported in Yivi (only back in IRMA), that does not work.
So now we just make one inner conjunction for each credential.