Stop using clientReturnUrl
On mobile, this redirects to a new tab, so it only works if we can restore
the full state of the frontend there. That's quite some work and might not
even be better, so I won't do it for now.
On android, without return URL it should work as expected (going back to
the app = browser that opened yivi), and on iOS using the back button to
the browser is the best we can do at least for now.