Make message+signature a base64 SignedMessage in verify url fragment
This has many advantages:
- There's just one blob, so it's less error-prone to copy-paste.
- For the user it's clear that the blob is meaningless if they don't verify it.
  After all, only by doing so can they read the message content.
- Very nice UX: the signature can be opened as a URL, and they'll be
  taken directly to the verification result. No need to copy-paste at all.
  This also means that the signature blob more or less documents itself:
  even without a header that describes how and where to verify it,
  the fact that it looks like (is) a URL, hints at opening it.

The length of the fragment does not seem problematic. On Firefox it works perfectly.
The internet suggests that other browsers probably don't limit it either, because the
fragment is of course not sent to any server. _Maybe_ Chrome would limit it, but if
it does, likely to 2MB, which is much larger than a signature would be.